+++ src/index.ts
@@ -1,23 +1,14 @@
-export { assembleHangul } from './assemble';
-export { chosungIncludes } from './chosungIncludes';
-export { choseongIncludes } from './choseongIncludes';
-export { combineHangulCharacter, combineVowels, curriedCombineHangulCharacter } from './combineHangulCharacter';
-export { convertQwertyToHangul, convertQwertyToHangulAlphabet } from './convertQwertyToHangulAlphabet';
-export { disassembleHangul, disassembleHangulToGroups } from './disassemble';
-export { disassembleCompleteHangulCharacter } from './disassembleCompleteHangulCharacter';
-export { hangulIncludes } from './hangulIncludes';
+export { assemble } from './assemble';
+export { combineCharacter, combineVowels } from './combineCharacter';
+export { convertQwertyToHangul, convertQwertyToAlphabet } from './convertQwertyToAlphabet';
+export { disassemble, disassembleToGroups } from './disassemble';
+export { disassembleCompleteCharacter } from './disassembleCompleteCharacter';
 export { josa } from './josa';
-export { removeLastHangulCharacter } from './removeLastHangulCharacter';
-export {
-  canBeChosung,
-  canBeJongsung,
-  canBeJungsung,
-  getChosung,
-  hasBatchim,
-  hasProperty,
-  hasSingleBatchim,
-  hasValueInReadOnlyStringList,
-} from './utils';
-export { extractHangul } from './extractHangul';
-export { acronymizeHangul } from './acronymizeHangul';
+export { removeLastCharacter } from './removeLastCharacter';
+export { romanize } from './romanize';
+export { standardizePronunciation } from './standardizePronunciation';
 export { susa } from './susa';
+export { hasBatchim } from './hasBatchim';
+export { canBeChoseong, canBeJongseong, canBeJungseong } from './canBe';
+export { getChoseong } from './getChoseong';
+export { amountToHangul } from './amountToHangul';